Created: exceptions handler and custom exceptions for varios bad requests

+++ src/main/java/com/addressbook/model/AddressBookData.java
@@ -12,7 +12,7 @@ public class AddressBookData {
 	private String address;
 	private String city;
 	private String state;
-	private long zip;
+	private int zip;
 	private long phoneNumber;
 	private String email;
